// frontend/src/shared/api/generated/models/TonNetwork.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger TON Lottery
 * TON Lottery
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: almostinf@mail.ru
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TonNetwork = {
    TonNetworkTestnet: 'TON_NETWORK_TESTNET',
    TonNetworkMainnet: 'TON_NETWORK_MAINNET'
} as const;
export type TonNetwork = typeof TonNetwork[keyof typeof TonNetwork];


export function instanceOfTonNetwork(value: any): boolean {
    for (const key in TonNetwork) {
        if (Object.prototype.hasOwnProperty.call(TonNetwork, key)) {
            if (TonNetwork[key as keyof typeof TonNetwork] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TonNetworkFromJSON(json: any): TonNetwork {
    return TonNetworkFromJSONTyped(json, false);
}

export function TonNetworkFromJSONTyped(json: any, ignoreDiscriminator: boolean): TonNetwork {
    return json as TonNetwork;
}

export function TonNetworkToJSON(value?: TonNetwork | null): any {
    return value as any;
}

export function TonNetworkToJSONTyped(value: any, ignoreDiscriminator: boolean): TonNetwork {
    return value as TonNetwork;
}